clean up the sources-tree tests
